requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# BigQuery / SQL
google-cloud-bigquery>=3.11.0
sqlalchemy>=2.0.0
pyarrow>=12.0.0

# Bayesian modeling
pymc>=5.7.0
arviz>=0.16.0
pytensor>=2.17.0

# Data processing
pydantic>=2.0.0
python-dateutil>=2.8.0

# Airflow
apache-airflow>=2.8.0
apache-airflow-providers-google>=10.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0
